fix: resolve SonarCloud gate findings and reviewer-flagged latent bugs
Addresses the open SonarCloud Quality Gate conditions (Security D, Reliability E
on new code) and the genuine bugs from the Copilot/github-code-quality review
threads. Where the finding was a real bug the production code was fixed; trusted-
path / unreachable false positives are left as-is and answered on their threads.

Security (Sonar java:S5443 - publicly writable temp dir, CRITICAL):
- PDFEncryptionUtil: write the encrypted PDF via PathValidationUtils.createSecure-
  TempFile (OWNER-only perms) instead of Files.createTempFile in the world-readable
  system temp dir; drop the now-unused java.nio.file.Files import.

Reliability:
- HL7A04Data: wrap FileWriter/BufferedWriter in try-with-resources so the writer is
  closed even if write() throws (BLOCKER - resource leak).
- MeasurementManager.getFlowsheetDsHTML: null-check File#listFiles() results and the
  ClassLoader#getResource() URL before dereferencing (NPE when a configured/classpath
  flowsheet dir is missing or unreadable).
- EDocUtil: pass the Integer ticklerNo straight to getTickler(...) instead of
  ticklerNo.intValue() (avoids an unbox that the Integer overload makes pointless).
- CaseManagementPrint: act on File#delete()'s return for temp print PDFs (log on
  failure) rather than ignoring it.

EForm import correctness:
- EFormExportZip: when an image already exists the code logged "skipping image" but
  still overwrote it - skip the write (else branch) so behaviour matches the message;
  drop a dead `html == null` check (html is `new String(...)`, never null).

Test hygiene:
- ZipUtilTest: add the CARLOS GPL/copyright header per the new-file standard.

Compiles clean (checkstyle included); 213 tests across the touched areas pass.

+++ src/main/java/io/github/carlos_emr/carlos/documentManager/EDocUtil.java
@@ -1119,7 +1119,7 @@ public final class EDocUtil {
         if (linkList != null) {
             for (TicklerLink tl : linkList) {
                 ticklerNo = tl.getTicklerNo();
-                Tickler t = ticklerManager().getTickler(loggedInInfo, ticklerNo.intValue());
+                Tickler t = ticklerManager().getTickler(loggedInInfo, ticklerNo);
                 HtmlTickler += "<br>" + Encode.forHtml(t.getMessage());
             }
         }
